style: 메인페이지 모바일 레이아웃 수정
