+++ virtualFittingApp/src/widgets/main/ui/AITextSection.tsx
@@ -115,7 +115,7 @@ const ScreenText = styled.div`
   mix-blend-mode: difference;
 
   @media (max-width: ${BREAKPOINTS.md}px) {
-    top: 3vh;
+    top: 10vh;
     left: 50%;
     font-size: clamp(60px, 15vw, 100px);
     transform: translate(-50%, 0);
@@ -136,6 +136,7 @@ const Wrapper = styled.div`
   @media (max-width: ${BREAKPOINTS.md}px) {
     align-items: center;
     padding-left: 0;
+    justify-content: flex-start;
   }
 `;
 
@@ -157,6 +158,7 @@ const Content = styled.div`
     margin-right: 0;
     width: 90%;
     align-items: center;
+    margin-top: 20vh;
   }
 `;
