+++ virtualFittingApp/src/widgets/main/ui/AboutSection.tsx
@@ -33,7 +33,7 @@ function AboutSection() {
         opacity: 1,
         y: 0,
         duration: 0.5,
-        stagger: 0.08,
+        stagger: 0.12,
         ease: "power2.out"
     });
 
@@ -50,8 +50,8 @@ function AboutSection() {
         masterTl.fromTo(
             textElement, 
             { opacity: 0, y: 20 },
-            { opacity: 1, y: 0, duration: 0.8, ease: "power2.out" },
-            "-=0.8"
+            { opacity: 1, y: 0, duration: 1.5, ease: "power2.out" },
+            ""
         );
     }
