+++ virtualFittingApp/src/widgets/main/ui/ServiceTextSection.tsx
@@ -117,7 +117,7 @@ const ScreenText = styled.div`
   mix-blend-mode: difference;
 
   @media (max-width: ${BREAKPOINTS.md}px) {
-    top: 3vh;
+    top: 10vh;
     left: 50%;
     font-size: clamp(60px, 15vw, 100px);
     transform: translate(-50%, 0);
@@ -138,6 +138,7 @@ const Wrapper = styled.div`
   @media (max-width: ${BREAKPOINTS.md}px) {
     align-items: center;
     padding-right: 0;
+    justify-content: flex-start;
   }
 `;
 
@@ -159,6 +160,7 @@ const Content = styled.div`
     margin-right: 0;
     width: 90%;
     align-items: center;
+    margin-top: 20vh;
   }
 `;
