+++ virtualFittingApp/src/widgets/main/ui/SolutionSection.tsx
@@ -109,7 +109,7 @@ function SolutionSection() {
 
 const Wrapper = styled.div`
   width: 100%;
-  height: auto;
+  height: 100%;
   display: flex;
   flex-direction: column;
   justify-content: center;
